Break import cycles by splitting getCollectin.js from collections.js
