+++ packages/vulcan-lib/lib/modules/collections.js
@@ -7,6 +7,8 @@ import { getSetting, registerSetting } from './settings.js';
 import { registerFragment, getDefaultFragmentText } from './fragments.js';
 import escapeStringRegexp from 'escape-string-regexp';
 import { validateIntlField, getIntlString, isIntlField } from './intl';
+import { getCollection, Collections } from './getCollection.js';
+export * from './getCollection.js';
 
 const wrapAsync = Meteor.wrapAsync ? Meteor.wrapAsync : Meteor._wrapAsync;
 // import { debug } from './debug.js';
@@ -27,14 +29,6 @@ export const viewFieldAllowAny = {allowAny:true};
 // will be set to `true` if there is one or more intl schema fields
 let hasIntlFields = false;
 export const getHasIntlFields = () => hasIntlFields
-
-export const Collections = [];
-
-export const getCollection = name =>
-  Collections.find(
-    ({ options: { collectionName } }) =>
-      name === collectionName || name === collectionName.toLowerCase()
-  );
 
 // TODO: find more reliable way to get collection name from type name?
 export const getCollectionName = typeName => Utils.pluralize(typeName);
