+++ packages/vulcan-lib/lib/modules/getCollection.js
@@ -1,0 +1,8 @@
+
+export const Collections = [];
+
+export const getCollection = name =>
+  Collections.find(
+    ({ options: { collectionName } }) =>
+      name === collectionName || name === collectionName.toLowerCase()
+  );
